requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
watchdog
Pillow
